+++ xarray/tests/test_plot.py
@@ -619,6 +619,8 @@ class Common2dMixin:
 
     def test_3d_raises_valueerror(self):
         a = DataArray(easy_array((2, 3, 4)))
+        if self.plotfunc.__name__ == 'imshow':
+            pytest.skip()
         with raises_regex(ValueError, r'DataArray must be 2d'):
             self.plotfunc(a)
 
@@ -669,6 +671,11 @@ class Common2dMixin:
     def test_can_plot_axis_size_one(self):
         if self.plotfunc.__name__ not in ('contour', 'contourf'):
             self.plotfunc(DataArray(np.ones((1, 1))))
+
+    def test_disallows_rgb_arg(self):
+        with pytest.raises(ValueError):
+            # Always invalid for most plots.  Invalid for imshow with 2D data.
+            self.plotfunc(DataArray(np.ones((2, 2))), rgb='not None')
 
     def test_viridis_cmap(self):
         cmap_name = self.plotmethod(cmap='viridis').get_cmap().name
@@ -1061,6 +1068,52 @@ class TestImshow(Common2dMixin, PlotTestCase):
     def test_2d_coord_names(self):
         with raises_regex(ValueError, 'requires 1D coordinates'):
             self.plotmethod(x='x2d', y='y2d')
+
+    def test_plot_rgb_image(self):
+        DataArray(
+            easy_array((10, 15, 3), start=0),
+            dims=['y', 'x', 'band'],
+        ).plot.imshow()
+        self.assertEqual(0, len(find_possible_colorbars()))
+
+    def test_plot_rgb_image_explicit(self):
+        DataArray(
+            easy_array((10, 15, 3), start=0),
+            dims=['y', 'x', 'band'],
+        ).plot.imshow(y='y', x='x', rgb='band')
+        self.assertEqual(0, len(find_possible_colorbars()))
+
+    def test_plot_rgb_faceted(self):
+        DataArray(
+            easy_array((2, 2, 10, 15, 3), start=0),
+            dims=['a', 'b', 'y', 'x', 'band'],
+        ).plot.imshow(row='a', col='b')
+        self.assertEqual(0, len(find_possible_colorbars()))
+
+    def test_plot_rgba_image_transposed(self):
+        # We can handle the color axis being in any position
+        DataArray(
+            easy_array((4, 10, 15), start=0),
+            dims=['band', 'y', 'x'],
+        ).plot.imshow()
+
+    def test_warns_ambigious_dim(self):
+        arr = DataArray(easy_array((3, 3, 3)), dims=['y', 'x', 'band'])
+        with pytest.warns(UserWarning):
+            arr.plot.imshow()
+        # but doesn't warn if dimensions specified
+        arr.plot.imshow(rgb='band')
+        arr.plot.imshow(x='x', y='y')
+
+    def test_rgb_errors_too_many_dims(self):
+        arr = DataArray(easy_array((3, 3, 3, 3)), dims=['y', 'x', 'z', 'band'])
+        with pytest.raises(ValueError):
+            arr.plot.imshow(rgb='band')
+
+    def test_rgb_errors_bad_dim_sizes(self):
+        arr = DataArray(easy_array((5, 5, 5)), dims=['y', 'x', 'band'])
+        with pytest.raises(ValueError):
+            arr.plot.imshow(rgb='band')
 
 
 class TestFacetGrid(PlotTestCase):